+++ app/Controllers/Api/SuperAdminApi.php
@@ -453,7 +453,7 @@ class SuperAdminApi extends BaseApiController
             'acceptanceLimitDays'  => (float) getSystemSetting('offer_acceptance_limit_days', 7),
             'ratingPeriod'         => (float) getSystemSetting('seller_rating_period_days', 7),
             'rejectionWindowHours' => (float) getSystemSetting('seller_rejection_window_hours', 24),
-            'minRentalDays'        => (int) getSystemSetting('min_rental_days', 3),
+            'minRentalDays'        => (float) getSystemSetting('min_rental_days', 3),
         ]);
     }
